Build Linux binaries in an old CentOS 5 image
This should help the binaries be maximally compatible with older Linux releases
as they'll depend on a very old glibc. This also touches up the linkage to
openssl to be static on Linux and dynamic on OSX.

The docker image used here is currently the same one that we're using in
rust-lang/rust for the production of the nightly compilers.