Re-implement and test Manifestation::update
This is the function that keeps an installation updated on
a distribution channel.

`update` takes a manifest describing a release of Rust
(which may be either a freshly-downloaded one, or the
same one used for the previous install), as well as lists
off extension components to add and remove.

From that it schedules a list of components to uninstall
and to uninstall to bring the installation up to date.
It downloads the components' packages. Then in a Transaction
uninstalls old packages and installs new packages,
writes the distribution manifest to "rustlib/multirust-dist.toml"
and a configuration containing the component name-target
pairs to "rustlib/multirust-config.toml".

This implementation is simpler than the existing one,
mostly because it doesn't reuse the manifest to represent
what is installed, but instead stores a list of installed
components.

Move MockInstallerBuilder into rust-install.