Add a wrapper type to allow converting errors to Box<Error>
This makes it easier to use multirust and associated crates as
libraries. This does not implement `error::Error` directly on the error
types to avoid having to almost-duplicated the `Dispaly::fmt()`
implementation in `Error::description()`. Instead, it adds a wrapper
that uses the `Display` impl to create and store the description. For
more discussion, see
  https://github.com/Diggsey/multirust-rs/pull/60