Produce more builds of rustup on Travis
This commit adds a few new targets for the rustup binary to follow suit with
Cargo and the compiler. This should help go that "last mile" in running rustc on
these platforms hopefully!

* `arm-unknown-linux-gnueabi`
* `arm-unknown-linux-gnueabihf`
* `armv7-unknown-linux-gnueabihf`
* `x86_64-unknown-freebsd`

All builds happen entirely within the standard `rust-slave-linux-cross`
container that Cargo and the compiler use, and all builds of rustup now compiler
OpenSSL locally as well (to follow suit with Cargo).