Tweak the command verbs and establish 'synonyms'
This reorganizes the verbs used in rustup commands. The concept here
is that every subcommand uses the most appropriate 'add/remove' verb
pair for itself, but implements hidden alternate commands using
the obvious other verbs.

So now the problematic verbs are:

* `toolchain install` / `toolchain uninstall`
* `target add` / `target remove`
* `override set` / `override unset`

Additional (hidden) synonyms are:

* `toolchain update`
* `toolchain add` / `toolchain remove`
* `target add` / `target remove`

This leaves the top level `rustup update` as the only non-hidden
'update' command, which has the unique focus of updating everything.