Catch potential misuses of target triples in host triple position
Before executing the `default` and `toolchain add` actions, check if the
toolchain given has a different triple than host. If so, bail out with a
hint explaining the subcommands and suggesting `target add` instead.

This is based on the assumption that one can't execute binaries from a
different triple; however there are niche scenarios where it is actually
possible, by QEMU user emulation and similar means. We might add a `-f`
switch for forcing the old behavior if it turns out to be necessary for
the relevant parties.

Fixes #826.
Fixes #834.