Fix the self-update tests to use the correct version of the bins
I realized that, even in the installer tests, most of the binaries
being run are out of the clitest 'exedir', not from CARGO_HOME/bin.
So this patch fixes that by modifying the generic test runner
to run bins off of the PATH, with PATH configured by default prefixed
with 'exedir'. The cli-self-upd tests extend PATH with CARGO_HOME/bin.