Test that previously-downloaded files are found and re-used
The test is only that the file-reuse notification is fired. In order to
detect this in the test, I've pushed through the notification handler
that gets given to `manifestation.update`, which in practice is nearly
always a clone of the one in DownloadCfg (but notably not when called as
in the test cases).