Switch on travis's caching features.
We can re-use the following across builds:
* compiled OpenSSL libs
* compiled crate dependencies

Use specific folder caching because we've got a custom installation
of rust/cargo etc. so travis's `cargo` caching can't really be
expected to get that right. Move the install location of
OpenSSL to ensure that if the version is bumped, we won't get a bad
cache hit and silently fail to apply the upgrade.

If the rust toolchain gets bumped, it would be worth checking what
gains can be got by incremental compilation of the local crate.