Consistently give toolchain argument a help text
This adds a help text to every toolchain command line argument. The
help text for 'rustup default' now reads:

  rustup-default
  Set the default toolchain

  USAGE:
      rustup default <toolchain>

  FLAGS:
      -h, --help    Prints help information

  ARGS:
      <toolchain>    Toolchain name, such as 'stable', 'nightly', or '1.8.0'.
                     For more information see `rustup help toolchain`

  DISCUSSION:
      Sets the default toolchain to the one specified. If the toolchain
      is not already installed then it is installed first.

Two commands (update and install) already explained what the toolchain
argument was -- but it was explained in the freeform "after help"
text. These explanations have been moved to the arguments themselves
for consistency.

Fixes #923.