Catch some missing components that weren't warned on
There is a lot of refactoring here too.

There was a check to ensure that if there are components that we plan to install (or update) that are missing from the manifest, then we error out. However, this only worked for components that had an entry in the manifest under the Rust package, but were missing components of their own. For tools like the RLS, when they are missing, they are completely missing from the manifest. They therefore didn't get picked up by the check. This patch adds an extra Vec to track such components and then warns about them.