Use `openssl-src` from crates.io to link to OpenSSL
This commit updates the CI builds for rustup to always build
`openssl-src` from crates.io instead of having a baked-in build of
OpenSSL. This matches what Cargo has switched to as well, and is
intended to centralize OpenSSL updates to one crates.io crate.

This also modernizes the docker containers and builds a bit to remove
the need for some scripts here and there.