Unify home directory handling.
The `std::env::home_dir()` function is deprecated (as of 1.29.0)
due to potentially "surprising" behaviour.  The compiler recommends
we switch to `dirs::home_dir()` so this commit does just that.

In addition, we risked diverging in behaviour because rustup-utils
has its own `home_dir()` behaviour for Windows, so change rustup-cli
to use that too.