Replace Windows custom toolchain hack
When using a custom toolchain on Windows, Rustup ensures that the
fallback Cargo calls the correct rustc executable by hardlinking cargo
into an empty directory and executing it there. The problem with this
method is that if cargo is invoked twice at the same time, the second
invocation fails because it can't replace the hardlink which it is in use
by the first. This occurs when an integration test itself invokes cargo.

This commit replaces the hack by specifying the correct version of rustc
by setting the RUSTC environment variable when cargo is invoked. This is
set to the absolute path of rustc in the correct toolchain rather than
the rustup wrapper. This is consistent with the way rustc is invoked on
a normal toolchain on Windows.