Version: Update git-testament to 0.1.4 and trust `stable`
In order that our CI process can continue to work, we need to
have a slightly newer version of `git-testament` which supports
indicating that the stable branch was the one used to build the
release binaries and so should trust the package version.

This also updates git-testament-derive to 0.1.5 which is more
reliable that 0.1.4 in the face of multiple branches at the same
commit.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>