download_tracker.rs: Erase line more robustly
In #1781, we switched to using `delete_line()` but unfortunately
that turns out to not work very well on Windows.  So this commit
turns us back to using spaces to erase things, but in a way which
ought to be moderately more robust than before.  There's still
no guarantee we're doing it right, but at least this time if
the terminal is approximately wide enough for the bar and no more,
we won't be flowing onto the next line with loads of spaces after
the rest of the tracker.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>