Add a PowerShell implementation for rustup-init
This change adds a new PowerShell script (`rustup-init.ps1`) which
performs the same duties and behaviors as the existing Shell-based
script (`rustup-init.sh`).

It is primarily targeted at tier 1 supported Windows platforms, namely:

* `i686-pc-windows-gnu` (32-bit MinGW)
* `i686-pc-windows-msvc` (32-bit MSVC)
* `x86_64-pc-windows-gnu` (64-bit MinGW)
* `x86_64-pc-windows-msvc` (64-bit MSVC)

However, due to the recent portability of PowerShell releases on macOS
and Linux, this script also supports these targets which overlap with
PowerShell targets:

* `x86_64-apple-darwin` (64-bit OS X/macOS)
* `x86_64-unknown-linux-gnu` (64-bit Linux)

As much as possible, the behavior of the existing `rustup-init.sh` is
preserved, including:

* Accepting arguments to be passed through to
  `rustup.init`/`rustup-init.exe`
* Short-circuiting the `-h` and `--help` flags to display an inline
  usage
* Honor the `RUSTUP_UPDATE_ROOT` environment variable if set, and
  defaulting to `https://static.rust-lang.org/rustup`
* Attempting to force TLS v1.2 when downloading the executable
* Downloading executable to a dedicated temporary directory under the
  system's temp path which is cleaned up on sucess and failure

For example, to install the latest nightly toolchain as the default in
an unattended scenario, a user can run the following in a PowerShell
session:

```powershell
.\rustup-init.ps1 -y --default-toolchain nightly
```

If `cmd.exe` is your current shell, you can still invoke the same by
running the above in a PowerShell session with:

```batch
powershell -Command ".\rustup-init.ps1 -y --default-toolchain nightly"
```

The script is written carefully to not call `exit` so that it can be
downloaded and used in much the same way as the Shell based "curl-bash"
strategy. Assuming that this script was available at
`https://ps1.rustup.rs` (or `https://pwsh.rustup.rs`,
`https://posh.rustup.rs`, etc.), and that no further arguments are
passed to the script, a user could invoke this script directly in a
PowerShell session with:

```powershell
iex ((New-Object System.Net.WebClient).DownloadString('https://ps1.rustup.rs'))
```
(For folks new to PowerShell, `iex` is an alias for the
`Invoke-Expression` cmdlet, similar to `eval` in other languages)

In order to pass arguments to the script, a slightly longer form can be
used. For example, to perform an unattended install of Rustup with no
default toolchain and no path modification, a user could run the
following in a PowerShell session:

```powershell
& ([scriptblock]::Create((New-Object System.Net.WebClient).DownloadString('https://ps1.rustup.rs'))) -y --default-toolchain none --no-modify-path
```

Yes, that's rather long, but I think that's as short a "code golf" as
you can get and it's on par with Chocolatey's `install.ps1` strategy.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>