Document (some) caveats when developing Rust on Windows
This commit documents errors that a developer might face when compiling Rust projects on Windows 10.

It is by far not a comprehensive list of errors and solutions, but its a start, and document the errors I've faced when trying to compile https://github.com/rust-lang/www.rust-lang.org on Windows 10, on two scenarios:

- Fresh `rustup` installation, following the instructions on screen, and installing VS Build Tools 2019
- Older installation of `rustup`, using VS Build Tools 2017

It is important to also be aware that following the instructions of installing VS Build Tools 2019 might require patches on `*-sys` crates, as their `build.rs` might not work for the new locations on the new edition of MS Build Tools. For example: https://github.com/compass-rs/sass-rs/pull/49