clitools: Hopefully mitigate unexpected test failures
In order to mitigate the test failures we were seeing where we
ran out of hard links on NTFS filesystems, copy the rustup binary
into place in the test environment, rather than hard-linking
the one which was built.  In order to do this safely, we use
an RwLock to ensure we can't be forking in other tests while
one test is copying a binary into place.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>