utils:  Fix cargo home test
In order to isolate this test from the test environment, if it
is attempting to determine behaviour when the CARGO_HOME environment
variable isn't present, it should ensure it's not present.

In additon, we need to test against \.cargo for Windows and we need
to allow for the `C:` in the second half of the test.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>