Support for machine wide installs: fallback-settings file
This provides a way for snaps, distro packages and other such installs
to set the default-toolchain for users that have not gone through the
rustup-init one-time set of questions.

/etc/rustup/settings.toml can more now just provide a default toolchain
for users of rustup on the machine. More keys will be supported as
needed. No equivalent file is defined for Windows OS machines at this
time.