Clarify help message with the Cargo home directory
Before this PR, the help message is:

```
Cargo's bin directory, located at:

        {cargo_home_bin}

This can be modified with the CARGO_HOME environment variable.
```

above **WILL** potentially mislead the user to set `CARGO_HOME` to `{cargo_home_bin}`, which is wrong.

This PR clearify the concept of `the Cargo home directory`, to separate from `Cargo's bin directory`, in fact, it's 2 different concepts and folders.