Fix the shebang for rustup-init.sh
The [normal installation guidance](https://rustup.rs/) is to pipe the contents of this file into `sh`, which is not `bash` (it may be `dash` or `bash` in POSIX-mode in certain systems, neither of which support most Bash-isms). Given that this script is intended to be run by `sh` the shebang should reflect that.