markdown: Fix indentation of multi-line code blocks
Before this commit the "Welcome to Rust!" text was:

```
This path will then be added to your PATH environment variable by
modifying the profile files located at:

  /Users/alexander/.profile
/Users/alexander/.zprofile
```

Notice the second line of the code block is not indented.
This is because the `Text` is passed to us as a single
`&str` (which contains newlines). We need to split the
text and call `LineWrapper.write_indent()` for each line.