Abstract over the current process
In order to do in-process testing process state - the current pid, args,
environment variables, cwd, and well known file descriptors and process
exit need to be abstracted out.

This patch introduces a system for doing so that is low (but not zero)
overhead, while relatively unintrusive.

A side effect of not calling exit() from deep in the process was to fix
a bug where downloads and tmp directory cleanup was not happening when
self updates were downloaded.