Document "invalid channel name '[toolchain]'" error
It took me a while to figure this out. I got this error when building a
project with rust-toolchain containing TOML and I had a fresh rustup,
rust nightly etc. but still there were errors. Running

    rustup default stable
    rustup default nightly

made it work for me. I'm not entirely sure if the explanation is totally
accurate, of course.