(profiles): Add support for `empty` profile
In order to support the possibility that you need a specific
tool from a specific `nightly`, we permit the installation of
an empty profile.  To do this, we have to pass permission for
the core manifest installer to perform a "no-op" which we only
ever pass in if you are installing a toolchain and you specify
that you want `--profile empty`.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>