Rephrase system-wide vs user-specific bash completion file dir note
It's not that much about whether a command is system-wide or
user-specific, but more about whether the completion is.

In particular, the user-specific completions file dir is used to
override system-wide completion, or to add completion for a command that
does not have a system-wide completion installed. This is orthogonal to
whether the command is installed system-wide or user-specific.