clippy::needless_borrow: Apply clippy suggestions
`cargo clippy` suggests removal of borrows in a number of locations.
Seems like the standard library has changed to allow for borrows to be
omitted on the result of `format!` and in a few other cases.

https://rust-lang.github.io/rust-clippy/master/index.html#needless_borrow

`CONTRIBUTING.md` suggests one runs this, before publishing a change:
```bash
cargo clippy --all --all-targets --all-features -- -D warnings`
```
But, at the moment, this command generated more than a thousand warnings
for me on `master`.

This particular warning is responsible for about 50 cases.

This change is just `cargo clippy --fix`, with flags to only show
`clippy::needless_borrow`, followed by `cargo fmt`.