Add in opentelemetry tracing as a feature
This is an evolution on the previous tracing support I added a few years
back. See CONTRIBUTING for details.

The rs_tracing crate is functionally fine but very limited in features
compared to the tracing crate that provides the developer interface for
opentelemetry instrumentation. As we start looking at performance again
I think this tooling will be useful.

Included is support for both tracing rustup-init and individual tests