Add support for listing components with target omitted
Configuration management like Chef or Ansible commonly checks this list
to know whether a component is present, with a view to installing it if
it is not.

This is slightly unwieldy at present because while installing can use an
implicit target (e.g., `rustup component add foo`), target-specific
components always show their target when listing (e.g.,
`foo-x86_64-unknown-linux-gnu`). The naive way to tackle this is to just
strip after the first dash, but of course this doesn't work well with
things like rust-analyzer.

There are other ways to achieve this (e.g., split from the end, since
the target is always a triple, although that can also get tricky when
the component is target agnostic), but it would be nice to just have a
--no-target flag which permits just showing the component short name for
tools which just want to simply check if the component is installed or
not, and don't really care about anything other than the implied target.