Fix rustup-init failing to read ZDOTDIR from zsh when SHELL is not zsh
If the SHELL environment variable does not include "zsh", rustup-init
attempts to call zsh to see if ZDOTDIR is set.

It intends to call
    zsh -c 'echo $ZDOTDIR'
but actually called
    zsh -c "'echo $ZDOTDIR'"

The second form fails with this output to stderr:
    zsh:1: command not found: echo $ZDOTDIR

Additionally, the default "echo" output includes a newline, and
this was being added to the PathBuf.  This would cause later
is_file() checks to fail, ignoring the rcfile.