Upload release artifacts to new S3 bucket
We are refactoring the release process for Rustup, which includes a new
S3 bucket for build artifacts. In the new release process, every commit
to the default branch builds and uploads artifacts. When a new release
is cut, these pre-built artifacts get copied into the current S3 bucket
for further testing and distribution.

The artifacts are currently uploaded to both the old and new bucket to
maintain backwards compatibility while implementing the new process.