Use a dedicated S3 bucket for Rustup
Based on a discussion[^1] with the release team, the decision has been
made to create a new S3 bucket for Rustup build artifacts. This enable
us to revoke the access that Rustup's CI has to the release bucket,
improving security and reducing the risk of malicious or accidental
compromise of Rustup releases.

[^1]: https://github.com/rust-lang/rustup/pull/3844#discussion_r1617167846