Have mocked cargo better adhere to cargo conventions
Currently, when the mocked cargo command is passed
`--recursive-cargo-subcommand`, it runs:

```
cargo-foo --recursive-cargo
```

However, cargo subcommands are normally passed their subcommand name as
the first argument. Thus, one would expect the following to be run:

```
cargo-foo foo --recursive-cargo
```

This commit changes the mocked cargo command to do the latter. It also
adds a check to ensure that the "subcommand name as first argument"
behavior is respected.

This is, admittedly, a rather pedantic change.

Extracted from #4175