style: partially migrate away from `if-let`
Following
<https://users.rust-lang.org/t/rust-2024-compatibility-lint-and-the-if-let-temporary-scope/125969/5>,
I have manually audited all `if_let_rescope` cases and found out that
certain existing cases can be rewritten to make it clearer semantically.
So these migrations have been made in advance so that the only cases of
`if-let` remaining are false positives of this lint.

The crux of determining false positives here is that, according to
<https://github.com/rust-lang/rust/issues/133167>, `if let Some()`
and other similar cases where the `else` drop is not significant are
not affected by the Edition change.