Switch flate2 to use the zlib-rs backend
This backend is written in Rust, but has substantially higher
performance. And, as of the latest version, zlib-rs does not
export any C zlib symbols, so it won't conflict with other
implementations of zlib in the same process; thus, it doesn't conflict
with linking to C libraries that use zlib.