rustup check: set exit status based on available updates
This helps with interfaces above rustup, for example prompting a user to
update, but only if updates are actually available.

If any updates are available, exits 0, otherwise exits 1.  This allows
"natural" shell commands like 'if rustup check; then <update>'.  It also
mirrors existing exit codes for update, where rustup exits 1 if updates
are not permitted by the distribution or because of errors.