feat(custom-toolchains): `target` and `component list` working on custom toolchains
As we cannot assume the existence of a manifest for custom toolchains,
we can only report the installed components/targets.
Thus, running `rustup target list` or `rustup target list --installed`,
will yield the same result.

The `list_items()` function was also refactored so that both custom and
distributable toolchains can make use of it.