feat(updates): check for updates concurrently
This change implied the introduction of the `futures_util` crate.
Feedback is greately appreciated on the above.

To ensure that there are no hangs, a special "if" was needed to ensure
that we never try running `.buffered(0)`. Otherwise, this would cause an
hang (kudos to rami3l for helping me out with this).

To report progress in a concurrent way whilst maintaining order, the
`indicatf` crate was used.
Note that, if we are not in a *tty* the progress wil be hidden and will
instead be printed after every channels is checked for updates.

NB: since we are using `indicatif` to display progress, when running in
a *tty*, we can use `buffer_unordered` instead of `buffered`, as
`indicatif` already handles the output in the correct order.
A quick benchmark (100 runs + 10 warmpu runs in a 50Mbps connection)
showed that this small change yields a **1.2x** speedup.