fix(downloads): honor the RUSTUP_CONCURRENT_DOWNLOADS by always having "n" concurrent downloads
Using this environment variable with a value between 2 and 5 would mean
that concurrency was not being totally maximized as there could be less
than n futures running at a some point in time.
See:
https://docs.rs/futures-util/0.3.31/futures_util/stream/trait.StreamExt.html#method.buffered

Adding a semaphore allows US to control how many futures are running at
a time, fixing this problem and ensuring that the env var is always
honored and the downloads are maximizing concurrency.