Specify --speed-limit and --speed-time for the fetch script
This change will cause curl to cancel and retry the connection
if the connection to the remote is exceedingly slow.

Note this change does apply some "policy" by expecting the user's
internet connection to be at least 20% the speed of modern DSL:
If the connection can't sustain at least 250,000 bytes per second
for 15 seconds, curl will disconnect and try again.

I don't know if the rust project has data on users in that
threshold or not, but this is upstreaming a change from our fork
of this script in DeterminateSystems/nix-installer.