fix(downloads): correct faulty output when retrying a download
The `DownloadFinished` notification was always being called even if the
download was unsuccessful. This would confuse the DownloadTracker as to
the download really ended (successfully).

To fix this, the notification is now only thrown iff the download didn't
timeout. This could be (possibly) accompanied by a new notification, say
`DownloadTimedOut` but it wouldn't have any direct use yet, so I refrain
from adding it.