fix(list): Match show command's styling
This changes the parenthetical from `bold` to `CONTEXT` to match what
was done in #4556.

This does not remove the bolding of the item itself.
We could do that. I figured its fine for now since there is a clearer
meaning with `installed` than when you can have `active` and `default`
on separate items.