fix(toolchain): Have 'list' match 'show's styling
`rustup show` and `rustup toolchain list` have similar lists but the
latter was unstyled.
This updates it to be styled to match `rustup show` as of #4556.

Maybe there would be a way to single source this in the future.
They do handle paths differently, so it would likely require a larger
conversation.