Add common commands section in help text
Judging from user feedback coming out of Debian
distros, it appears that after installing `rustup`
from the package manager it is not immediately
clear which command installs the toolchain.
This patch adds a small text so that after
installation of `rustup`, which normally does not
invoke the toolchain installation for the user,
the user can still receive guidance on how to set
up the local toolchain immediately.

Signed-off-by: Xiangfei Ding <dingxiangfei2009@protonmail.ch>